[android]: video and audio cleanup
Finishing touches on video and audio,
handling lifecycle events and fixing
inconsistencies.

+++ ViroRenderer/VROVideoTexture.h
@@ -45,7 +45,7 @@ public:
     virtual void pause() = 0;
     virtual void play() = 0;
     virtual bool isPaused() = 0;
-    virtual void seekToTime(int seconds) = 0;
+    virtual void seekToTime(float seconds) = 0;
 
     virtual void setMuted(bool muted) = 0;
     virtual void setVolume(float volume) = 0;
